Added bones to dino

+++ RobGL/RobGL/VertexWeightData.h
@@ -1,10 +1,11 @@
 #pragma once
 #include <External/glm/common.hpp>
+#include <External/glm/gtc/type_precision.hpp>
 namespace rgl {
 	struct VertexWeightData
 	{
 	public:
-		glm::ivec4 Joints;
+		glm::u16vec4 Joints;
 		glm::vec4 Weights;
 	};
 }
